Move EnsureTillerInstalled to separate file
